third_party/WebKit/LayoutTests/fast/css/resize-corner-tracking-touch-expected.txt

Test for https://code.google.com/p/chromium/issues/detail?id=150601 Can't resize a textarea with touch.

a placeholder so that we have enough elements to scroll the page


PASS successfullyParsed is true

TEST COMPLETE


Resize div at offset (-6,-7)
PASS target.offsetWidth - old_width is deltaXExpected
PASS target.offsetHeight - old_height is deltaYExpected

Resize div at offset (6,7)
PASS target.offsetWidth - old_width is deltaXExpected
PASS target.offsetHeight - old_height is deltaYExpected

Resize textarea1 at offset (-20,-20)
PASS target.offsetWidth - old_width is deltaXExpected
PASS target.offsetHeight - old_height is deltaYExpected

Resize iframe1 at offset (-6,-7)
PASS target.offsetWidth - old_width is deltaXExpected
PASS target.offsetHeight - old_height is deltaYExpected

Resize iframe1 at offset (-20,-20)
PASS target.offsetWidth - old_width is deltaXExpected
PASS target.offsetHeight - old_height is deltaYExpected

Resize textarea2in iframe iframe1 at offset (-6,-7)
PASS target.offsetWidth - old_width is deltaXExpected
PASS target.offsetHeight - old_height is deltaYExpected
